requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
beautifulsoup4==4.12.3
mammoth==1.10.0
python-docx==1.1.2
lxml==5.3.0
